requirements.txt
Flask==0.12
gunicorn==19.3.0
Flask-Jsonpify
google-api-python-client
google-cloud-firestore
firebase-admin
grpcio
requests
pytz